fix(usb):change uintptr_t to uint64_t

+++ model/usb/host/include/usb_pnp_notify.h
@@ -62,7 +62,7 @@ struct UsbInfoQueryPara {
     UsbInfoQueryParaType type;
     union {
         int32_t id;
-        uintptr_t usbDevAddr;
+        uint64_t usbDevAddr;
         struct {
             int32_t devNum;
             int32_t busNum;
